MF-000 Move MetroMap under /map
